Add exp, exp2, exp10 and expm1

+++ src/common/misc.h
@@ -172,7 +172,7 @@ typedef struct {
 #if defined(ENABLEFLOAT128)
 typedef __float128 Sleef_quad;
 #else
-typedef struct { uint64_t x, y; } Sleef_quad;
+typedef struct { double x, y; } Sleef_quad;
 #endif
 #endif
